Users table qualities added
The add_qualities_to_users.rb file was created, which defines the changes made to the users table by adding columns. The references[] for matches quality, I commented out as it was giving me an error. Check the database via sqlworkbench and check the migration file in the migrate directory under db. Also added a template.txt file as we previously made edits to the schema file. This template file holds all the qualities we want to implement into the tables, and is located in the db directory.