Fixed null:false error
Seems to be that by setting a column of the user with null:false, will result in an error. Fixed it by removing this contraint on the columns. Once we implement variables in the users model, we can then return some of the columns to null:false. Such as first name and last name, as the error handle will take care of no input for first or last name, meaning its never null and therefore never returning false.