Implemented User Home Routing
Implemented a change where routes.rb considers both authenticated users and unauthenticated users to determine their root path. If a user is authenticated, their root path is the home page, otherwise an unauthenticated user's root path is the login page. All unauthenticated users trying to access any page that isnt the login page, will be redirected to the login page and will be prompted with the according error message. Also implemented a couple buttons, My Profile and Logout, that are displayed at the top of every page if the user is signed in. This will be later implemented as a menu bar.