Show Profile and edit profile implementation
Implemented working pages of viewing current users profile as well as editing own profile. Fields for interests and college location are present and working, however, the design may want to be reconsidered to work alongside a library of interests and colleges accordingly. A select option could be useful as well if we decide to consider specific dorm areas as the college location. Cancel account is also working. Removed initial edit and update pages made with the install of the users controller as they were redundant to our structure.