Profile picture
Implemented profile picture functionality for the user class. Added according columns to the users table and used the paperclip gem. Thumb icon and medium icon are available. Other models can be customized as well. Users can upload/change their profile picture.   I put the thumb icon on the drop down menu for the user, and styled it briefly to give it a round border. This will need some minor tweaking on the frontend as it slightly distorts rectangular images when converting to a circular one. If the user does not have an image yet, the drop down menu button just displays "My Account".