Edit Activity Implemented & Removed Password Check
Edit activity routing is complete, and buttons for editting are displayed if the given activity project owner id is the current user. Does not display like and dislike buttons as well. After update, user is sent back to home. As this is all done through rendering the same form, for a new activity and editting an activity, the logic for checking if it is a create button or update button is also established. Similarly, depending on the form, either a new activity or editted one, the according image is displayed. Also made a method for the user to override the update parameters of devise. The user is now able to update every aspect of their account (besides new password), without a password prompt. New password is in the same form, but obviously left outside of the card. Needs to be styled as a expandable/collapsable part of the form in case user wants to edit their current password.