Activity Delete Functionality
Activity functionality for delete has been implemented. User can delete a given activity if they are project owner. Confirmation with "Are you sure?" message needs to be handled. Deleting activity also deletes all corresponding memberships. Please note that currently, the pages controller has a home method that pulls a specific activity by an id we manually put in. If the owner of this activity decides to delete this activity, the home page will be looking for an activity that no longer exists. Currently working on a hot fix to pull and pass an existing activity to the home page.