Created image methods for user and activity
Changed directory of default images such that all the default pictures, including the dropzone image, were routed properly. Also transformed the local image_tag calls for paperclip into calls to methods of the user/activity class accordingly. These methods are in the user and activity helper files. All images returning an image, take 3 arguments. The first argument is the object of the image, either activity or user. The 2nd argument is the size, thumb or medium, and the last argument is the class in which the image is styled.