Added team count column to activities table
Added capacity function that enables the activity card view to display the current amount of members over the max amount of members allowed. Minor styling to a button look, backend implementation is done, just needs to be considered for queue in concern to whether or not a full activity should be in a user queue.